refactoring: update folders structure
