+++ src/containers/ProductDetails/ProductDetails.js
@@ -1,11 +1,11 @@
 import React from 'react'
-import Screen from '../../styledComponents/Screen'
-import Button from '../../styledComponents/Button'
-import products from '../../data/products'
-import categories from '../../data/categories'
+import Screen from '../../components/Screen'
+import Button from '../../components/Button'
+import products from '../../lib/data/products'
+import categories from '../../lib/data/categories'
 import {connect} from 'react-redux'
 import { Link } from 'react-router-dom'
-import {addToCart} from '../../screens/ShoppingCart/actionCreators'
+import {addToCart} from '..//ShoppingCart/actionCreators'
 
 const mapDispatchToProps = (dispatch) => {
   return {
