+++ src/containers/ShoppingCart/actionTypes.js
