+++ src/index.js
@@ -1,7 +1,7 @@
 import 'core-js/fn/object/assign'
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './components/App/App'
+import App from './App/App'
 import { BrowserRouter } from 'react-router-dom'
 import store from './store'
 import {Provider} from 'react-redux'
