+++ src/lib/data/categories.js
