+++ src/lib/images/yeoman.png
